+++ lib/src/home/view/home_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:sozzle/src/common/border_elevated_button.dart';
+import 'package:sozzle/src/home/view/play_button.dart';
 import 'package:sozzle/src/theme/theme.dart';
 import 'package:sozzle/src/user_stats/user_stats.dart';
 
@@ -12,8 +13,8 @@ class HomePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return BlocBuilder<ThemeCubit, ThemeState>(
-      builder: (context, state) => Scaffold(
-        backgroundColor: state.backgroundColor,
+      builder: (context, theme) => Scaffold(
+        backgroundColor: theme.backgroundColor,
         body: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
@@ -21,50 +22,32 @@ class HomePage extends StatelessWidget {
               child: Text(
                 'Sozzle Home Page',
                 style: TextStyle(
-                  color: state.primaryTextColor,
+                  color: theme.primaryTextColor,
                   fontSize: 24,
                 ),
               ),
             ),
+            const SizedBox(
+              height: 20,
+            ),
             BlocBuilder<UserStatsCubit, UserStatsState>(
               builder: (context, state) {
-                return TextButton(
-                  onPressed: () {
-                    // TODO(any): route to game play
-                    context.read<UserStatsCubit>().advanceLevelUp();
-                  },
-                  child: Text('Current Level: ${state.progress.currentLevel}'),
-                );
+                return PlayButton(state.progress.currentLevel.toString());
               },
             ),
             const SizedBox(
               height: 20,
             ),
             BorderElevatedButton(
-              route: '/home/settings',
+              route: '/settings',
               text: 'Settings',
-              backgroundColor: Colors.blue,
-              borderColor: Colors.white,
-              textColor: state.primaryTextColor,
+              backgroundColor: theme.backgroundColor,
+              borderColor: theme.primaryTextColor,
+              textColor: theme.primaryTextColor,
             ),
           ],
         ),
       ),
-    );
-  }
-}
-
-class PlayButton extends StatelessWidget {
-  const PlayButton(this.levelId, {super.key});
-  final String levelId;
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: () {
-        GoRouter.of(context).replace('/play/$levelId');
-      },
-      child: Text('PLAY Level: $levelId'),
     );
   }
 }
